[appconfig] move to shared rollup config from dev-tool
while on this, also remove circular dependency of helpers => featureFlag |
secretReference => helper. There are only two calls to the helper, worth
duplicating to remove the circular dependency rollup warning.
