+++ sdk/appconfiguration/app-configuration/rollup.config.js
@@ -1,13 +1,6 @@
-import * as base from "./rollup.base.config";
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
 
-const inputs = [];
+import { makeConfig } from "@azure/dev-tool/shared-config/rollup";
 
-if (!process.env.ONLY_BROWSER) {
-  inputs.push(base.nodeConfig());
-}
-
-if (!process.env.ONLY_NODE) {
-  inputs.push(base.browserConfig());
-}
-
-export default inputs;
+export default makeConfig(require("./package.json"));
